% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{check_names}
\usage{
check_names(x, data)
}
\arguments{
\item{x}{a vector of colnames.}

\item{data}{a data.frame or imputationList.}
}
\value{
check_names will trigger an error if the supplied vector of names were not found in the supplied object. It will also predict which names the user was trying to spell.
}
\description{
Takes in a vector of colnames, and a data.frame or imputationList. This function will trigger an error if names are not in the data object. In addition, this function will try to predict which names the user was trying to spell.
}
