% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_similar.R
\name{find_similar}
\alias{find_similar}
\title{find_similar}
\usage{
find_similar(x, y, threshold = 4)
}
\arguments{
\item{x}{a vector of strings.}

\item{y}{a vector of strings. Strings will be returned from y which are similar to those in x}

\item{threshold}{a numeric. Strings in y will be returned if they are less than this distance away from x}
}
\value{
a vector containing strings from y, which are similar to those in x.
}
\description{
Takes in two vectors of strings. This function will return strings in the second set, which are similar to those in the first.
}
