% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.numeric.R
\name{get.numeric}
\alias{get.numeric}
\title{get.numeric}
\usage{
get.numeric(x)
}
\arguments{
\item{x}{a data.frame or 'imputationList'}
}
\value{
Returns the original object with all non-numeric columns removed.
}
\description{
Takes in a data.frame or imputationList, removes non-numeric columns, and returns the object.
}
