% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_names.R
\name{check_names}
\alias{check_names}
\title{check_names}
\usage{
check_names(x, data)
}
\arguments{
\item{x}{a vector of names.}

\item{data}{a data.frame or imputationList.}
}
\value{
check_names will trigger an error if the supplied vector of names are not found in the supplied object. It will also predict which names the user was trying to spell.
}
\description{
Takes in a vector of colnames, and a data.frame or imputationList. This funciton will trigger an error if names are not in the data object. In addition, this name will try to predict which names they were trying to spell.
}
